Add -Wconversion and fix all warnings
-Wconversion warns if an integer conversion may
change sign or value. There were many such warnings,
fix all of them by changing to more appropriate integer
types or, where safe, use explicit casts.

+++ meminfo.h
@@ -33,7 +33,7 @@ void print_mem_stats(int (*out_func)(const char* fmt, ...), const meminfo_t m);
 int get_oom_score(int pid);
 int get_oom_score_adj(const int pid, int* out);
 long long get_vm_rss_kib(int pid);
-int get_comm(int pid, char* out, int outlen);
+int get_comm(int pid, char* out, size_t outlen);
 int get_uid(int pid);
 
 #endif
